feat(dashboard): ajout top 10 clients par chiffre d'affaires
- Endpoint GET /api/dashboard/top-clients
- TopClientDto avec companyName et revenue
- Méthode getTopClients dans DashboardService
- Tests unitaires
- Requête SQL optimisée avec agrégation

+++ backend/src/main/java/com/gp_dev/erp_lite/dtos/TopClientDto.java
@@ -14,5 +14,5 @@ public class TopClientDto {
     private Long clientId;
     private String clientName;
     private BigDecimal totalRevenue;
-    private Integer invoiceCount;
+    private Long invoiceCount;
 }
